Source code/Backend/requirements.txt
Flask>=2.0
tensorflow==2.19.0
pillow>=9.0
numpy>=1.23
werkzeug>=2.0
